# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIKARIR API",
        "description": "Dokumentasi API untuk aplikasi Sikarir — Sistem Informasi Karir & Pelatihan BLK",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/sertifikat": {
            "get": {
                "tags": [
                    "Sertifikat"
                ],
                "summary": "Ambil semua data sertifikat (paginate 15)",
                "operationId": "66af7636ad36139e8c8104818bcfff1d",
                "responses": {
                    "200": {
                        "description": "Daftar sertifikat berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sertifikat"
                ],
                "summary": "Upload sertifikat baru untuk peserta",
                "operationId": "1399d8ea58dd9ba616f295f7c1ad264b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "peserta_id",
                                    "jadwal_id"
                                ],
                                "properties": {
                                    "peserta_id": {
                                        "type": "integer"
                                    },
                                    "jadwal_id": {
                                        "type": "integer"
                                    },
                                    "admin_id": {
                                        "type": "integer"
                                    },
                                    "nomor_sertifikat": {
                                        "type": "string"
                                    },
                                    "tanggal_terbit": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "file_sertifikat": {
                                        "description": "File sertifikat (PDF/gambar)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sertifikat berhasil diupload"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sertifikat/{id}": {
            "get": {
                "tags": [
                    "Sertifikat"
                ],
                "summary": "Ambil detail sertifikat berdasarkan ID",
                "operationId": "fe7104f3a1afcd3c5d9bc6212bcda5d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sertifikat"
                    },
                    "404": {
                        "description": "Sertifikat tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sertifikat"
                ],
                "summary": "Update data sertifikat (gunakan POST dengan _method=PUT untuk upload file)",
                "operationId": "d598b5b78634b0b1faa1388bca767ffa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nomor_sertifikat": {
                                        "type": "string"
                                    },
                                    "tanggal_terbit": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "file_sertifikat": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "default": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sertifikat berhasil diupdate"
                    },
                    "404": {
                        "description": "Sertifikat tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sertifikat"
                ],
                "summary": "Hapus data sertifikat beserta filenya",
                "operationId": "c791989f656615261d59529065d9f6eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sertifikat berhasil dihapus"
                    },
                    "404": {
                        "description": "Sertifikat tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token yang diperoleh dari endpoint /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}